refactor: 카운트다운 관련 로직 커스텀 훅 분리 #270
